# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Blink Shop API
  version: 0.12.0
  description: |
    Blink Shop 的单一接口契约。约定：
    - 所有路径前缀为 `/api/v1`，JSON 字段使用 snake_case。
    - 成功响应直接返回业务对象，不包 `data` 外层。
    - 失败响应统一为 `Error`；状态码含义见 `docs/03-api-contract.md`。
    - 每个响应都带 `X-Request-ID`；请求可携带 `X-Request-ID`（1–64 位字母、数字、`.`、`_`、`-`），否则由服务端生成。
    - 请求体上限：JSON 默认 1 MiB；单个上传文件默认 10 MiB（`UPLOAD_MAX_BYTES`），multipart 请求体另留 64 KiB 给表单分隔符；超限返回 413。
    - 未登录请求按客户端 IP 限流，已登录请求另按账号限流，超限返回 429 并带 `Retry-After`。
    - 非流式请求默认 30 秒超时，超时返回 504。
    - 未知路径返回 404，路径存在但方法不支持返回 405（带 `Allow` 头），均为 `Error` JSON。
    - 认证：`Authorization: Bearer <token>`。token 由注册/登录返回，服务端只保存 SHA-256 摘要，默认 24 小时过期（`AUTH_TOKEN_TTL`），登出即撤销。
      token 缺失返回 401 `unauthorized`（请先登录）；过期、撤销或伪造同样 401 `unauthorized`（登录已失效）。
    - 角色：路由分为 public / session / account / user / merchant / admin 六类，角色不符返回 403 `forbidden`；
      商家只能操作自己店铺的资源，越权 403。完整矩阵见 backend/README.md。
    - 账户状态：inactive 只能调用 `GET /auth/me` 与 `POST /auth/logout`，其余 403 `account_inactive`；
      risk 可以读（GET）和登出，任何写操作 403 `account_risk`；两者都不能新登录。已注销账户的 token 全部失效。
    - 分页参数 `page` 从 1 开始，`page_size` 默认 10、最大 100；列表响应使用 `Page` 结构。
      参数缺省或不是整数时取默认值，超出范围时取边界值（与上游一致，不报 400），响应中的 `page`/`page_size` 为实际使用的值。
    - 金额是两位小数的字符串（如 `"2999.00"`），比例是四位小数的字符串（如 `"0.9500"`）。
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
    description: 与当前访问地址同源（本地开发为 http://localhost:8080）
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [auth]
      operationId: register
      summary: 注册普通用户
      description: 只能注册 `user` 角色。成功后直接返回会话，等同于登录。用户名不区分大小写，已注销账户的用户名不可复用。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: 注册成功
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: 用户名已存在（`username_exists`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: username_exists
                message: 账号已存在
                request_id: 6f1c2a9e0b7d4c3e8a5f1b2c
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/login:
    post:
      tags: [auth]
      operationId: login
      summary: 用户名密码登录
      description: |
        用户不存在、已注销和密码错误统一返回 401 `invalid_credential`。密码正确但账户 inactive/risk 时返回 403
        `account_inactive` / `account_risk`。同一用户名每分钟最多尝试 `LOGIN_ATTEMPTS_PER_MINUTE` 次（默认 10），超出 429。
        旧版 SHA-256 密码哈希在登录成功后自动升级为 bcrypt。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 登录成功
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: 账号或密码错误（`invalid_credential`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: invalid_credential
                message: 账号或密码错误
                request_id: 6f1c2a9e0b7d4c3e8a5f1b2c
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /auth/logout:
    post:
      tags: [auth]
      operationId: logout
      summary: 登出
      description: 撤销本次请求使用的 token，同一账户的其他会话不受影响。inactive/risk 账户也可调用。
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: 当前登录账户
      description: inactive/risk 账户也可调用，客户端据 `status` 展示受限提示。
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /account/profile:
    get:
      tags: [account]
      operationId: getProfile
      summary: 读取个人资料
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
    patch:
      tags: [account]
      operationId: updateProfile
      summary: 修改昵称或头像
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /account/contact:
    patch:
      tags: [account]
      operationId: updateContact
      summary: 修改手机号和邮箱
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdateRequest"
      responses:
        "200":
          $ref: "#/components/responses/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /account:
    delete:
      tags: [account]
      operationId: deleteAccount
      summary: 注销账户
      description: 软删账户并撤销其全部 token；用户名不可再注册。inactive/risk 账户不能注销。
      responses:
        "200":
          $ref: "#/components/responses/OK"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
  /uploads/avatar:
    post:
      tags: [account]
      operationId: uploadAvatar
      summary: 上传头像
      description: |
        multipart 字段 `file`，最大 2MB；按文件内容判断类型，只接受 JPG、PNG、WebP。只返回地址，
        需要再调用 `PATCH /account/profile` 设置 `avatar_url`。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  contentMediaType: application/octet-stream
      responses:
        "200":
          description: 上传成功
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvatarUpload"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /uploads/avatar/{name}:
    get:
      tags: [account]
      operationId: getAvatar
      summary: 读取头像（公开）
      security: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            pattern: "^acct_[A-Za-z0-9_]{1,64}_[0-9a-f]{16}\\.(jpg|png|webp)$"
      responses:
        "200":
          description: 图片内容
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
        "404":
          $ref: "#/components/responses/NotFound"
  /files:
    post:
      tags: [files]
      operationId: uploadFile
      summary: 上传私有文件
      description: |
        multipart 字段 `file`（其他字段忽略）。服务端边接收边计算大小和 SHA-256，按文件内容嗅探类型，
        客户端声明的 Content-Type 和文件名一律忽略；默认允许 JPG、PNG、WebP、GIF、PDF
        （`UPLOAD_ALLOWED_MIME_TYPES`，只能在 JPG/PNG/WebP/GIF/BMP/PDF 中选择），单个文件默认最大 10 MiB（`UPLOAD_MAX_BYTES`）。
        文件保存到对象存储，key 随机生成且不对外返回；先写对象再写元数据，任一步失败都不会留下文件记录。
        所有已登录角色可用；inactive / risk 账户不能上传。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  contentMediaType: application/octet-stream
      responses:
        "201":
          description: 上传成功
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileUpload"
        "400":
          description: |
            `invalid_argument`：缺少 `file` 分段、文件为空或 multipart 格式错误（field=file 表示文件本身的问题）；
            `unsupported_file_type`：内容不是允许的类型（field=file）。
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: unsupported_file_type
                message: 不支持该文件类型，仅支持 JPG、PNG、WEBP、GIF、PDF
                field: file
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          $ref: "#/components/responses/ObjectStorageUnavailable"
  /files/{id}:
    get:
      tags: [files]
      operationId: getFile
      summary: 下载私有文件
      description: |
        只有上传者本人和管理员可以读取，其他账号 403。成功时直接返回文件字节：`Content-Type` 为上传时嗅探出的类型，
        带 `X-Content-Type-Options: nosniff`、`Content-Security-Policy: default-src 'none'; sandbox`，
        图片 `Content-Disposition: inline`，其他类型 `attachment`；`Cache-Control: private, max-age=3600`，
        `ETag` 为内容 SHA-256，带匹配的 `If-None-Match` 时返回 304。
        Web 端 `<img>` 无法携带 Bearer token，需要用带 token 的请求取回内容后生成 blob URL（见 frontend/src/api/files.ts），
        不能把 token 放进 URL。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^file_[A-Za-z0-9_]{1,64}$"
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: 文件内容
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            ETag:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
          content:
            image/jpeg: {}
            image/png: {}
            image/webp: {}
            image/gif: {}
            image/bmp: {}
            application/pdf: {}
        "304":
          description: 内容未变化
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: 不是上传者本人也不是管理员（`forbidden`），或账户已停用（`account_inactive`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: forbidden
                message: 无权访问该文件
        "404":
          description: 文件不存在（`file_not_found`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: file_not_found
                message: 文件不存在
        "503":
          $ref: "#/components/responses/ObjectStorageUnavailable"
  /search/image:
    post:
      tags: [files]
      operationId: searchImage
      summary: 按图片找商品
      description: |
        先用 `POST /files` 上传图片，再用返回的 `file_id` 检索。只能用本人上传的图片：别人的文件与不存在的文件一样返回 404。
        服务端从对象存储读取图片、解码校验（JPG/PNG/WebP/GIF/BMP，单张不超过 10 MiB、边长 8–8000 像素），生成图片向量，
        在商品图索引里找最相似的商品，命中一律回到商品库按可见性取回——下架、风控、删除的商品和停业店铺的商品不会返回。
        `status`：`matched` 最像的商品达到同款阈值；`similar` 只有外观相近的商品（只作参考）；`no_match` 没有相似商品
        （包括纯色、几乎空白的图片）。相似度只反映外形和配色，价格、库存以商品详情为准。
        没有配置图片搜索（未配置 Milvus）时返回 503 `image_search_unavailable`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_id]
              properties:
                file_id:
                  type: string
                  pattern: "^file_[A-Za-z0-9_]{1,64}$"
                top_k:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 6
      responses:
        "200":
          description: 检索结果
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageSearchResult"
        "400":
          description: |
            `invalid_argument`：缺少或不合法的 `file_id`、`top_k` 不在 1–20；`unsupported_file_type`：文件不是图片（field=file_id）；
            `invalid_image`：图片无法解码或尺寸超限（field=file_id）。
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: invalid_image
                message: 图片无法识别，请换一张清晰的 JPG / PNG 图片
                field: file_id
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "404":
          description: 文件不存在或不是本人上传的（`file_not_found`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: file_not_found
                message: 文件不存在
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "503":
          description: 图片搜索未配置或索引不可用（`image_search_unavailable`），或对象存储不可用（`object_storage_unavailable`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: image_search_unavailable
                message: 图片搜索暂不可用，请稍后再试或用文字描述商品
  /speech/tts/config:
    get:
      tags: [speech]
      operationId: getSpeechConfig
      summary: 语音能力配置
      description: |
        返回语音合成和实时识别是否开通、供应商名、默认音色和限制，客户端据此决定是否显示朗读 / 语音输入按钮。
        不包含任何凭据。仅普通用户可用。
      responses:
        "200":
          description: 能力配置
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpeechConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
  /speech/tts:
    post:
      tags: [speech]
      operationId: synthesizeSpeech
      summary: 朗读文本（语音合成）
      description: |
        把文本合成音频，成功直接返回音频字节（`Content-Type` 为 `audio/mpeg`、`audio/wav` 等，`Cache-Control: no-store`）。
        客户端应先清洗文本（去掉 Markdown、链接、商品 ID 等），长度不超过能力配置里的 `max_text_chars`。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: 去掉首尾空白后 1–max_text_chars 个字
                voice:
                  type: string
                  pattern: "^[A-Za-z0-9_-]{1,64}$"
                  description: 音色；为空用默认音色
      responses:
        "200":
          description: 音频字节
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            audio/mpeg: {}
            audio/wav: {}
            audio/ogg: {}
        "400":
          description: "`invalid_argument`：文本为空或超长（field=text）、音色名不合法（field=voice）"
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: invalid_argument
                message: 请提供要朗读的文字
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/AccountStatusForbidden"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "501":
          description: 没有配置语音合成（`tts_not_enabled`）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: tts_not_enabled
                message: 语音朗读暂未开通
        "502":
          description: 供应商出错或超时（`tts_failed`）；详细原因只记在服务端日志（不含凭据）
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: tts_failed
                message: 语音合成服务暂时不可用，请稍后再试
...
  /admin/configs:
    get:
      tags: [admin]
      operationId: listAdminConfigs
      summary: 配置列表
      description: 全部已注册的配置项（不分页），密钥只返回掩码。值的优先级为环境变量 → 动态配置 → 默认值。
      responses:
        "200":
          description: 配置项
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConfigItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 登录后获得的不透明 token。
  headers:
    RequestID:
      description: 请求追踪 ID，与错误响应体中的 `request_id` 一致。
      schema:
        type: string
        pattern: "^[A-Za-z0-9._-]{1,64}$"
...
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
...